@@ -1,12 +1,16 @@
-import * as React from "react";
+import * as React from 'react';
 
-import "./style.scss";
+import './style.scss';
 
 const Footer = () => (
-    <footer>
-        <span className="footer-copyright">© 2018 Parseclabs. All Rights Reserved.</span>
-        <a className="footer-email" href="mailto:info@parseclabs.org">info@parseclabs.org</a>
-    </footer>
+  <footer>
+    <span className="footer-copyright">
+      © 2018 Parseclabs. All Rights Reserved.
+    </span>
+    <a className="footer-email" href="mailto:info@parseclabs.org">
+      info@parseclabs.org
+    </a>
+  </footer>
 );
 
 export default Footer;
