@@ -1,14 +1,14 @@
-import * as React from "react";
+import * as React from 'react';
 
-import "./style.scss";
+import './style.scss';
 
 export default () => (
-    <div className="pulsating-dot">
-        <div className="pulsating-dot__ripple">
-            <span/>
-            <div/>
-            <div/>
-            <div/>
-        </div>
+  <div className="pulsating-dot">
+    <div className="pulsating-dot__ripple">
+      <span />
+      <div />
+      <div />
+      <div />
     </div>
+  </div>
 );
