@@ -1,94 +1,119 @@
-import "./style.scss";
+import './style.scss';
 
-import * as React from "react";
-import { observer } from "mobx-react";
-import Select, { OptionComponentProps, OptionValues } from "react-select";
+import * as React from 'react';
+import { observer } from 'mobx-react';
+import Select, { OptionComponentProps, OptionValues } from 'react-select';
 
-import { Token } from "../../../../types";
-import { fromCents } from "../../../../../../../../utils";
+import { Token } from '../../../../types';
+import { fromCents } from '../../../../../../../../utils';
 
-const dropdown = require("./img/dropdown.svg");
-const dropdownWhite = require("./img/dropdown-white.svg");
+const dropdown = require('./img/dropdown.svg');
+const dropdownWhite = require('./img/dropdown-white.svg');
 
 export type SelectTokenProps = {
-    selected: Token,
-    color: number,
-    balances: Array<any>,
-    tokens: Array<Token>
-    onChange: (token: Token) => void
+  selected: Token;
+  color: number;
+  balances: Array<any>;
+  tokens: Array<Token>;
+  onChange: (token: Token) => void;
 };
 
 export type SelectTokenState = {
-    selected: Token | null
+  selected: Token | null;
 };
 
-class Option extends React.PureComponent<OptionComponentProps<OptionValues>> {
-    handleMouseDown = (event) => {
-        event.preventDefault();
-        event.stopPropagation();
+class Option extends React.PureComponent<OptionComponentProps<Token>> {
+  handleMouseDown = (event: React.MouseEvent<HTMLDivElement>) => {
+    event.preventDefault();
+    event.stopPropagation();
 
-        this.props.onSelect(this.props.option, event);
+    if (this.props.onSelect) {
+      this.props.onSelect(this.props.option, event);
     }
+  };
 
-    render() {
-        const { option } = this.props;
-        const balance = fromCents(option.balance, option.decimals);
-        return (
-            <div className="alice-balance-panel_balance" onMouseDown={this.handleMouseDown}>
-                <img src={require(`./icons/${option.symbol}.svg`)} className="alice-balance-panel_balance-icon" />
-                <span>Balance: <strong>{balance} {option.symbol}</strong></span>
-            </div>
-        );
-    }
+  render() {
+    const { option } = this.props;
+    const balance = fromCents(option.balance, option.decimals);
+    return (
+      <div
+        className="alice-balance-panel_balance"
+        onMouseDown={this.handleMouseDown}
+      >
+        <img
+          src={require(`./icons/${option.symbol}.svg`)}
+          className="alice-balance-panel_balance-icon"
+        />
+        <span>
+          Balance:{' '}
+          <strong>
+            {balance} {option.symbol}
+          </strong>
+        </span>
+      </div>
+    );
+  }
 }
 
 @observer
-export default class SelectToken extends React.Component<SelectTokenProps, SelectTokenState> {
+export default class SelectToken extends React.Component<
+  SelectTokenProps,
+  SelectTokenState
+> {
+  static defaultProps = {
+    tokens: [],
+  };
 
-    static defaultProps = {
-        tokens: []
-    };
+  valueRenderer = (token: Token) => {
+    const balance = this.props.balances
+      ? this.props.balances[this.props.color]
+      : 0;
 
-    valueRenderer = (token) => {
-        const balance = this.props.balances ? this.props.balances[this.props.color] : 0;
-
-        return (
-            <div className="alice-balance-panel_balance">
-                <img src={require(`./icons/${token.symbol}.svg`)} className="alice-balance-panel_balance-icon" />
-                <span>Balance: <strong>{fromCents(balance, token.decimals)} {token.symbol}</strong></span>
-            </div>
-        );
-    }
-
-    arrowRenderer = ({ isOpen }) => (
+    return (
+      <div className="alice-balance-panel_balance">
         <img
-            className={`select-token-dropdown${isOpen ? " opened" : ""}`}
-            src={isOpen ? dropdownWhite : dropdown}
+          src={require(`./icons/${token.symbol}.svg`)}
+          className="alice-balance-panel_balance-icon"
         />
-    )
+        <span>
+          Balance:{' '}
+          <strong>
+            {fromCents(balance, token.decimals)} {token.symbol}
+          </strong>
+        </span>
+      </div>
+    );
+  };
 
-    handleSelectToken = (token: Token) => {
-        this.props.onChange(token);
-    }
+  arrowRenderer = ({ isOpen }: { isOpen: boolean }) => (
+    <img
+      className={`select-token-dropdown${isOpen ? ' opened' : ''}`}
+      src={isOpen ? dropdownWhite : dropdown}
+    />
+  );
 
-    render() {
-        const {tokens} = this.props;
-        return tokens.length > 0 ? (
-            <Select
-                className="select-token"
-                optionComponent={Option}
-                options={tokens.filter((t => t.symbol !== this.props.selected.symbol))}
-                placeholder="Select address"
-                onChange={this.handleSelectToken}
-                searchable={false}
-                multi={false}
-                value={this.props.selected}
-                valueRenderer={this.valueRenderer}
-                valueKey="account"
-                labelKey="name"
-                clearable={false}
-                arrowRenderer={this.arrowRenderer}
-            />
-        ) : null;
-    }
+  handleSelectToken = (token: Token) => {
+    this.props.onChange(token);
+  };
+
+  render() {
+    const { tokens } = this.props;
+    return tokens.length > 0 ? (
+      <Select
+        className="select-token"
+        optionComponent={Option}
+        options={tokens.filter(t => t.symbol !== this.props.selected.symbol)}
+        placeholder="Select address"
+        onChange={this.handleSelectToken}
+        searchable={false}
+        multi={false}
+        value={this.props.selected}
+        valueRenderer={this.valueRenderer}
+        valueKey="account"
+        labelKey="name"
+        clearable={false}
+        arrowRenderer={this.arrowRenderer}
+      />
+    ) : null;
+  }
 }
