@@ -1,9 +1,16 @@
-import * as React from "react";
+import * as React from 'react';
 
-import "./style.scss";
+import './style.scss';
 
-export default ({ onClick, disabled }) => (
-    <button className="tx-send" onClick={onClick} disabled={disabled}>
-        Send
-    </button>
+interface SendButtonProps {
+  onClick: React.MouseEventHandler;
+  disabled?: boolean;
+}
+
+const SendButton: React.SFC<SendButtonProps> = ({ onClick, disabled }) => (
+  <button className="tx-send" onClick={onClick} disabled={disabled}>
+    Send
+  </button>
 );
+
+export default SendButton;
