@@ -1,24 +1,26 @@
 export interface InputPanelProps {
-    balance: number;
-    onSend: Function;
+  balance: number;
+  onSend: Function;
+}
+
+export interface Receiver {
+  name: String;
+  account: String;
 }
 
 export interface InputPanelState {
-    amount: string | number | "";
-    receiver: null | {
-        name: String;
-        account: String;
-    };
-    sending: boolean;
-    sent: boolean;
-    hiddenNotification?: boolean;
-    notificationText?: string;
+  amount?: string | number | '';
+  receiver: null | Receiver;
+  sending: boolean;
+  sent: boolean;
+  hiddenNotification?: boolean;
+  notificationText?: string;
 }
 
 export interface SelectAddressOption {
-    onSelect: Function;
-    option: {
-        name: string;
-        account: string;
-    };
+  onSelect: Function;
+  option: {
+    name: string;
+    account: string;
+  };
 }
