@@ -1,5 +1,5 @@
 export type TransactionsPanelProps = {
-    transactions: Array<any>,
-    lastBlock: number,
-    getLastTransactions: any
+  transactions: Array<any>;
+  lastBlock: number;
+  getLastTransactions: any;
 };
