@@ -1,9 +1,6 @@
-import * as React from "react";
-import * as ReactDOM from "react-dom";
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
 
-import App from "./App";
+import App from './App';
 
-ReactDOM.render(
-  <App />,
-  document.getElementById("app") as HTMLElement
-);
+ReactDOM.render(<App />, document.getElementById('app') as HTMLElement);
