@@ -1,40 +1,49 @@
-import { observable, action } from "mobx";
-import { TransactionReceipt } from "web3/types";
-import getWeb3 from "../getWeb3";
+import { observable, action } from 'mobx';
+import { TransactionReceipt } from 'web3/types';
+import getWeb3 from '../getWeb3';
 
 export default class Transaction {
-    transactionHash: string;
-    @observable status: string | boolean;
-    @observable transactionIndex: number;
-    @observable blockHash: string;
-    @observable blockNumber: number;
-    @observable from: string;
-    @observable to: string;
-    @observable contractAddress: string;
-    @observable cumulativeGasUsed: number;
-    @observable gasUsed: number;
+  transactionHash: string;
+  @observable
+  status: string | boolean;
+  @observable
+  transactionIndex: number;
+  @observable
+  blockHash: string;
+  @observable
+  blockNumber: number;
+  @observable
+  from: string;
+  @observable
+  to: string;
+  @observable
+  contractAddress: string;
+  @observable
+  cumulativeGasUsed: number;
+  @observable
+  gasUsed: number;
 
-    constructor(receipt: TransactionReceipt) {
-        Object.assign(this, receipt);
-        this.loadReceipt(receipt.transactionHash);
-    }
+  constructor(receipt: TransactionReceipt) {
+    Object.assign(this, receipt);
+    this.loadReceipt(receipt.transactionHash);
+  }
 
-    @action
-    update = (receipt) => {
-        Object.assign(this, receipt);
-    }
+  @action
+  update = (receipt: any) => {
+    Object.assign(this, receipt);
+  };
 
-    @action
-    loadReceipt = (hash: string) => {
-        const web3 = getWeb3();
-        web3.eth.getTransactionReceipt(hash, (err, receipt) => {
-            if (err) {
-                console.error(err.message);
-            } else if (receipt) {
-                this.update({
-                    status: receipt.status
-                });
-            }
+  @action
+  loadReceipt = (hash: string) => {
+    const web3 = getWeb3();
+    web3.eth.getTransactionReceipt(hash, (err, receipt) => {
+      if (err) {
+        console.error(err.message);
+      } else if (receipt) {
+        this.update({
+          status: receipt.status,
         });
-    }
+      }
+    });
+  };
 }
