@@ -1,3 +1,3 @@
 export interface Web3Window extends Window {
-    web3: any;
+  web3: any;
 }
